refactor(host-infra-system): extract shared config store primitives
